Refactor SingletonKit、EasyEvent、ActionKit

+++ RunTime/Core/Architecture/IArchitecture.cs
@@ -100,7 +100,7 @@ namespace Framework3.Core
         /// <param name="onEvent">事件触发时的回调函数</param>
         /// <param name="priority">事件优先级</param>
         /// <returns>IUnRegister 接口，用于取消注册</returns>
-        IUnRegister RegisterEvent<TEvent>(Action<TEvent> onEvent, int priority);
+        IUnRegister RegisterEvent<TEvent>(Action<TEvent> onEvent, float priority);
 
         /// <summary>
         /// 注销事件监听
