+++ RunTime/Toolkits/Core/EventKit/EventTrigger/Mono/OnBecomeVisibleEventTrigger.cs
@@ -19,14 +19,14 @@ namespace Framework3.Toolkits.EventKit
 
     public static class OnBecameVisibleEventTriggerExtension
     {
-        public static IUnRegister OnBecameVisibleEvent<T>(this T self, Action onBecameVisible, int priority = 0)
+        public static IUnRegister OnBecameVisibleEvent<T>(this T self, Action onBecameVisible, float priority = 0)
             where T : Component
         {
             return self.GetOrAddComponent<OnBecomeVisibleEventTrigger>().OnBecameVisibleEvent
                        .Register(onBecameVisible, priority);
         }
 
-        public static IUnRegister OnBecameVisibleEvent(this GameObject self, Action onBecameVisible, int priority = 0)
+        public static IUnRegister OnBecameVisibleEvent(this GameObject self, Action onBecameVisible, float priority = 0)
         {
             return self.GetOrAddComponent<OnBecomeVisibleEventTrigger>().OnBecameVisibleEvent
                        .Register(onBecameVisible, priority);
