+++ RunTime/Toolkits/Core/EventKit/EventTrigger/Mono/OnDestroyEventTrigger.cs
@@ -19,14 +19,14 @@ namespace Framework3.Toolkits.EventKit
 
     public static class OnDestroyEventTriggerExtension
     {
-        public static IUnRegister OnDestroyEvent<T>(this T self, Action onDestroy, int priority = 0)
+        public static IUnRegister OnDestroyEvent<T>(this T self, Action onDestroy, float priority = 0)
             where T : Component
         {
             return self.GetOrAddComponent<OnDestroyEventTrigger>().OnDestroyEvent
                        .Register(onDestroy, priority);
         }
 
-        public static IUnRegister OnDestroyEvent(this GameObject self, Action onDestroy, int priority = 0)
+        public static IUnRegister OnDestroyEvent(this GameObject self, Action onDestroy, float priority = 0)
         {
             return self.GetOrAddComponent<OnDestroyEventTrigger>().OnDestroyEvent
                        .Register(onDestroy, priority);
