+++ RunTime/Toolkits/Core/EventKit/EventTrigger/Mono/OnUpdateEventTrigger.cs
@@ -17,14 +17,14 @@
 
     public static class OnUpdateEventTriggerExtension
     {
-        public static IUnRegister OnUpdateEvent<T>(this T self, Action update, int priority = 0)
+        public static IUnRegister OnUpdateEvent<T>(this T self, Action update, float priority = 0)
             where T : Component
         {
             return self.GetOrAddComponent<OnUpdateEventTrigger>().UpdateEvent
                        .Register(update, priority);
         }
 
-        public static IUnRegister OnUpdateEvent(this GameObject self, Action update, int priority = 0)
+        public static IUnRegister OnUpdateEvent(this GameObject self, Action update, float priority = 0)
         {
             return self.GetOrAddComponent<OnUpdateEventTrigger>().UpdateEvent
                        .Register(update, priority);
