+++ RunTime/Toolkits/Core/EventKit/EventTrigger/Physics/OnCollisionEnter2DEventTrigger.cs
@@ -27,7 +27,7 @@ namespace Framework3.Toolkits.EventKit
 
     public static class OnCollisionEnter2DEventTriggerExtension
     {
-        public static IUnRegister OnCollisionEnter2DEvent<T>(this T self, Action<Collision2D> onCollisionEnter2D, int priority = 0)
+        public static IUnRegister OnCollisionEnter2DEvent<T>(this T self, Action<Collision2D> onCollisionEnter2D, float priority = 0)
             where T : Component
         {
             return self.GetOrAddComponent<OnCollisionEnter2DEventTrigger>().OnCollisionEnter2DEvent
