+++ RunTime/Toolkits/Core/EventKit/EventTrigger/Physics/OnCollisionExit2DEventTrigger.cs
@@ -27,7 +27,7 @@ namespace Framework3.Toolkits.EventKit
 
     public static class OnCollisionExit2DEventTriggerExtension
     {
-        public static IUnRegister OnCollisionExit2DEvent<T>(this T self, Action<Collision2D> onCollisionExit2D, int priority = 0)
+        public static IUnRegister OnCollisionExit2DEvent<T>(this T self, Action<Collision2D> onCollisionExit2D, float priority = 0)
             where T : Component
         {
             return self.GetOrAddComponent<OnCollisionExit2DEventTrigger>().OnCollisionExit2DEvent
