+++ RunTime/Toolkits/Core/EventKit/EventTrigger/Physics/OnCollisionExitEventTrigger.cs
@@ -27,14 +27,14 @@ namespace Framework3.Toolkits.EventKit
 
     public static class OnCollisionExitEventTriggerExtension
     {
-        public static IUnRegister OnCollisionExitEvent<T>(this T self, Action<Collision> onCollisionExit, int priority = 0)
+        public static IUnRegister OnCollisionExitEvent<T>(this T self, Action<Collision> onCollisionExit, float priority = 0)
             where T : Component
         {
             return self.GetOrAddComponent<OnCollisionExitEventTrigger>().OnCollisionExitEvent
                        .Register(onCollisionExit, priority);
         }
 
-        public static IUnRegister OnCollisionExitEvent(this GameObject self, Action<Collision> onCollisionExit, int priority = 0)
+        public static IUnRegister OnCollisionExitEvent(this GameObject self, Action<Collision> onCollisionExit, float priority = 0)
         {
             return self.GetOrAddComponent<OnCollisionExitEventTrigger>().OnCollisionExitEvent
                        .Register(onCollisionExit, priority);
