+++ RunTime/Toolkits/Core/EventKit/EventTrigger/Physics/OnTriggerStay2DEventTrigger.cs
@@ -27,7 +27,7 @@ namespace Framework3.Toolkits.EventKit
 
     public static class OnTriggerStay2DEventTriggerExtension
     {
-        public static IUnRegister OnTriggerStay2DEvent<T>(this T self, Action<Collider2D> onTriggerStay2D, int priority = 0)
+        public static IUnRegister OnTriggerStay2DEvent<T>(this T self, Action<Collider2D> onTriggerStay2D, float priority = 0)
             where T : Component
         {
             return self.GetOrAddComponent<OnTriggerStay2DEventTrigger>().OnTriggerStay2DEvent
