+++ RunTime/Toolkits/Core/EventKit/EventTrigger/UI/OnBeginDragEventTrigger.cs
@@ -28,13 +28,13 @@ namespace Framework3.Toolkits.EventKit
 
     public static class OnBeginDragEventTriggerExtension
     {
-        public static IUnRegister OnBeginDragEvent<T>(this T self, Action<PointerEventData> onBeganDrag, int priority = 0)
+        public static IUnRegister OnBeginDragEvent<T>(this T self, Action<PointerEventData> onBeganDrag, float priority = 0)
             where T : Component
         {
             return self.GetOrAddComponent<OnBeginDragEventTrigger>().OnBeginDragEvent.Register(onBeganDrag, priority);
         }
 
-        public static IUnRegister OnBeginDragEvent(this GameObject self, Action<PointerEventData> onBeganDrag, int priority = 0)
+        public static IUnRegister OnBeginDragEvent(this GameObject self, Action<PointerEventData> onBeganDrag, float priority = 0)
         {
             return self.GetOrAddComponent<OnBeginDragEventTrigger>().OnBeginDragEvent.Register(onBeganDrag, priority);
         }
