+++ RunTime/Toolkits/Core/EventKit/EventTrigger/UI/OnCancelEventTrigger.cs
@@ -28,13 +28,13 @@ namespace Framework3.Toolkits.EventKit
 
     public static class OnCancelEventTriggerExtension
     {
-        public static IUnRegister OnCancelEvent<T>(this T self, Action<BaseEventData> onCancel, int priority = 0)
+        public static IUnRegister OnCancelEvent<T>(this T self, Action<BaseEventData> onCancel, float priority = 0)
             where T : Component
         {
             return self.GetOrAddComponent<OnCancelEventTrigger>().OnCancelEvent.Register(onCancel, priority);
         }
 
-        public static IUnRegister OnCancelEvent(this GameObject self, Action<BaseEventData> onCancel, int priority = 0)
+        public static IUnRegister OnCancelEvent(this GameObject self, Action<BaseEventData> onCancel, float priority = 0)
         {
             return self.GetOrAddComponent<OnCancelEventTrigger>().OnCancelEvent.Register(onCancel, priority);
         }
