+++ RunTime/Toolkits/Core/EventKit/EventTrigger/UI/OnDeselectEventTrigger.cs
@@ -29,13 +29,13 @@ namespace Framework3.Toolkits.EventKit
 
     public static class OnDeselectEventTriggerExtension
     {
-        public static IUnRegister OnDeselectEvent<T>(this T self, Action<BaseEventData> onDeselect, int priority = 0)
+        public static IUnRegister OnDeselectEvent<T>(this T self, Action<BaseEventData> onDeselect, float priority = 0)
             where T : Component
         {
             return self.GetOrAddComponent<OnDeselectEventTrigger>().OnDeselectEvent.Register(onDeselect, priority);
         }
 
-        public static IUnRegister OnDeselectEvent(this GameObject self, Action<BaseEventData> onDeselect, int priority = 0)
+        public static IUnRegister OnDeselectEvent(this GameObject self, Action<BaseEventData> onDeselect, float priority = 0)
         {
             return self.GetOrAddComponent<OnDeselectEventTrigger>().OnDeselectEvent.Register(onDeselect, priority);
         }
