+++ RunTime/Toolkits/Core/EventKit/EventTrigger/UI/OnEndDragEventTrigger.cs
@@ -28,13 +28,13 @@ namespace Framework3.Toolkits.EventKit
 
     public static class OnEndDragEventTriggerExtension
     {
-        public static IUnRegister OnEndDragEvent<T>(this T self, Action<PointerEventData> onEndDrag, int priority = 0)
+        public static IUnRegister OnEndDragEvent<T>(this T self, Action<PointerEventData> onEndDrag, float priority = 0)
             where T : Component
         {
             return self.GetOrAddComponent<OnEndDragEventTrigger>().OnEndDragEvent.Register(onEndDrag, priority);
         }
 
-        public static IUnRegister OnEndDragEvent(this GameObject self, Action<PointerEventData> onEndDrag, int priority = 0)
+        public static IUnRegister OnEndDragEvent(this GameObject self, Action<PointerEventData> onEndDrag, float priority = 0)
         {
             return self.GetOrAddComponent<OnEndDragEventTrigger>().OnEndDragEvent.Register(onEndDrag, priority);
         }
