+++ RunTime/Toolkits/Core/EventKit/EventTrigger/UI/OnMoveEventTrigger.cs
@@ -28,13 +28,13 @@ namespace Framework3.Toolkits.EventKit
 
     public static class OnMoveEventTriggerExtension
     {
-        public static IUnRegister OnMoveEvent<T>(this T self, Action<AxisEventData> onMove, int priority = 0)
+        public static IUnRegister OnMoveEvent<T>(this T self, Action<AxisEventData> onMove, float priority = 0)
             where T : Component
         {
             return self.GetOrAddComponent<OnMoveEventTrigger>().OnMoveEvent.Register(onMove, priority);
         }
 
-        public static IUnRegister OnMoveEvent(this GameObject self, Action<AxisEventData> onMove, int priority = 0)
+        public static IUnRegister OnMoveEvent(this GameObject self, Action<AxisEventData> onMove, float priority = 0)
         {
             return self.GetOrAddComponent<OnMoveEventTrigger>().OnMoveEvent.Register(onMove, priority);
         }
