+++ RunTime/Toolkits/Core/EventKit/EventTrigger/UI/OnPointerExitEventTrigger.cs
@@ -28,13 +28,13 @@ namespace Framework3.Toolkits.EventKit
 
     public static class OnPointerExitEventTriggerExtension
     {
-        public static IUnRegister OnPointerExitEvent<T>(this T self, Action<PointerEventData> onPointerExit, int priority = 0)
+        public static IUnRegister OnPointerExitEvent<T>(this T self, Action<PointerEventData> onPointerExit, float priority = 0)
             where T : Component
         {
             return self.GetOrAddComponent<OnPointerExitEventTrigger>().OnPointerExitEvent.Register(onPointerExit, priority);
         }
 
-        public static IUnRegister OnPointerExitEvent(this GameObject self, Action<PointerEventData> onPointerExit, int priority = 0)
+        public static IUnRegister OnPointerExitEvent(this GameObject self, Action<PointerEventData> onPointerExit, float priority = 0)
         {
             return self.GetOrAddComponent<OnPointerExitEventTrigger>().OnPointerExitEvent.Register(onPointerExit, priority);
         }
